+++ src/cpp/tardigrade_hydraPerzynaViscoplasticity.cpp
@@ -3258,6 +3258,29 @@ namespace tardigradeHydra{
 
         }
 
+        void residual::addParameterizationInfo( std::string &parameterization_info ){
+            /*!
+             * Add parameterization information to the incoming string
+             * 
+             * \param &parameterization_info: The parameterization info string
+             */
+
+            parameterization_info += "class: tardigradeHydra::perzynaViscoplasticity::residual\n\n";
+            parameterization_info += "name,                         description,       units, current value\n";
+            parameterization_info += "   n,        the Perzyna exponential term,        none, " + std::to_string(    ( *get_perzynaParameters( ) )[ 0 ] ) + "\n";
+            parameterization_info += "  q0,             the initial drag stress,      stress, " + std::to_string( ( *get_dragStressParameters( ) )[ 0 ] ) + "\n";
+            parameterization_info += "  q1, the drag stress hardening parameter,      stress, " + std::to_string( ( *get_dragStressParameters( ) )[ 1 ] ) + "\n";
+            parameterization_info += "  C1,                the WLF C1 parameter,        none, " + std::to_string(    ( *get_thermalParameters( ) )[ 0 ] ) + "\n";
+            parameterization_info += "  C2,                the WLF C2 parameter, temperature, " + std::to_string(    ( *get_thermalParameters( ) )[ 1 ] ) + "\n";
+            parameterization_info += "Tref,       the WLF reference temperature, temperature, " + std::to_string(    ( *get_thermalParameters( ) )[ 2 ] ) + "\n";
+            parameterization_info += "   Y,                initial yield stress,      stress, " + std::to_string(      ( *get_yieldParameters( ) )[ 0 ] ) + "\n";
+            parameterization_info += "   A,          yield pressure sensitivity,        none, " + std::to_string(      ( *get_yieldParameters( ) )[ 1 ] ) + "\n";
+            parameterization_info += "   B,           flow pressure sensitivity,        none, " + std::to_string(       ( *get_flowParameters( ) )[ 1 ] ) + "\n";
+            parameterization_info += " hi0,          isv initial evolution rate,        none, " + std::to_string(  ( *get_hardeningParameters( ) )[ 0 ] ) + "\n";
+            parameterization_info += " hi1,           isv linear evolution rate,        none, " + std::to_string(  ( *get_hardeningParameters( ) )[ 1 ] ) + "\n";
+
+        }
+
     }
 
 }