+++ docs/sphinx/changelog.rst
@@ -26,6 +26,7 @@ Breaking Changes
 Internal Changes
 ================
 - Added parameterization information for the perzynaJ2Viscoplasticity residual (:pull:`211`). By `Nathan Miller`_.
+- Added parameterization information for the perzynaViscoplasticity residual (:pull:`212`). By `Nathan Miller`_.
 
 ******************
 0.5.0 (05-22-2025)